+++ static-module/.gitignore
