+++ static-module/README.md
