+++ static-module/package.json
