--- static-module/school-journal/src/components/main/index.tsx
@@ -1,9 +1,0 @@
-import './main.scss'
-
-const Main = () => {
-    return (
-        <></>
-    )
-}
-
-export default Main
