+++ static-module/src/components/app/app.scss
