+++ static-module/src/index.scss
