+++ static-module/src/index.tsx
