+++ static-module/src/react-app-env.d.ts
