+++ static-module/src/typescript/types.ts
