+++ static-module/tsconfig.json
